+++ src/main/java/micycle/pgs/PGS_ShapePredicates.java
@@ -5,7 +5,9 @@ import static micycle.pgs.PGS_Conversion.fromPShape;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.vecmath.Point3d;
 import javax.vecmath.Point4d;
@@ -508,6 +510,45 @@ public final class PGS_ShapePredicates {
 			maxAngle = Math.max(maxAngle, Angle.interiorAngle(p0, p1, p2));
 		}
 		return maxAngle;
+	}
+
+	/**
+	 * Calculates the interior angles of a polygon represented by a {@link PShape}.
+	 * The method calculates the interior angle at each vertex.
+	 * <p>
+	 * The vertices of the input {@code shape} are assumed to represent a simple
+	 * polygon.
+	 *
+	 * @param shape The {@link PShape} representing the polygon for which to
+	 *              calculate interior angles. It's expected to be a polygon shape.
+	 * @return A map where keys are {@link PVector} vertices of the polygon and
+	 *         values are their corresponding interior angles in radians as double
+	 *         values.
+	 * @since 2.1
+	 */
+	public static Map<PVector, Double> interiorAngles(PShape shape) {
+		Map<PVector, Double> anglesMap = new HashMap<>();
+
+		var vertices = PGS_Conversion.toPVector(shape); // unclosed
+		if (!PGS.isClockwise(vertices)) {
+			Collections.reverse(vertices);
+		}
+
+		int n = vertices.size();
+		for (int i = 0; i < n; i++) {
+			PVector currentVertex = vertices.get(i);
+			PVector previousVertex = vertices.get((i - 1 + n) % n); // Get previous vertex, wrapping around
+			PVector nextVertex = vertices.get((i + 1) % n); // Get next vertex, wrapping around
+
+			Coordinate p0 = new Coordinate(previousVertex.x, previousVertex.y);
+			Coordinate p1 = new Coordinate(currentVertex.x, currentVertex.y);
+			Coordinate p2 = new Coordinate(nextVertex.x, nextVertex.y);
+			double angleRadians = Angle.interiorAngle(p0, p1, p2); // CW
+			anglesMap.put(currentVertex, angleRadians);
+		}
+
+		return anglesMap;
+
 	}
 
 	/**